Add files via upload
Content-based image retrieval is one of the interesting subjects in image processing and machine vision. In the image retrieval systems, the query image is compared with images in the database to retrieve images with similar contents. Image comparison will be done using extracted features from the query and database images. In this paper, the features are extracted based on the human visual system. Since the human visual system considers the texture and edge orientation in images for comparison, the colour difference histogram associated with texture and edge orientation in image is extracted as features. In this paper, the features are selected using the Shannon entropy criterion. The proposed method is tested using the Corel-5K and Corel-10K databases. The precision and recall criteria were used to evaluate the proposed system. The experimental results show the ability of the proposed system for more accurate retrieval rather than recently content-based image retrieval systems.